+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - refacto to have all the test files in the same place (for UT and IT), to avoid maintaining 2 test directories
 - refacto all test files to add sub-directories for each rule, to be more clear and to be able to add more tests for each rule in the future
 - fix integration test system run + fix TI GCI82
+- upgrade delivery process to be dynamic
 
 ## [2.1.2] - 2026-01-11