+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- compatibility updates for SonarQube up to 26.7.0
+
 ### Deleted
 
 ## [2.2.0] - 2026-06-16