+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - compatibility updates for SonarQube up to 26.7.0
 - update default CI JDK from 17 to 21
 - upgrade internal librairies versions
+- add CI check for JDK17 and JDK21
 
 ### Deleted